Refactored bonuses and removed unnecessary update() calls for static entities. Adjusted world dimensions to 500x800 (classic Doodle Jump size). Updated Main.cpp with tests for moving platforms, now functioning as expected. Tweaked velocity for better gameplay balance.
-- TODO: Improve collision checks and finalize world.h.

+++ src/Logic/Entities/Platforms/DisappearingPlatform.cpp
@@ -5,7 +5,7 @@
 #include "DisappearingPlatform.h"
 
 // DisappearingPlatform constructor
-DisappearingPlatform::DisappearingPlatform(int x, int y) : StaticPlatform(x, y) {
+DisappearingPlatform::DisappearingPlatform(float x, float y) : StaticPlatform(x, y) {
     disappearTime = 2.0f;
     elapsedTime = 0.0f;
     type = PlatformType::DISAPPEARING;
